+++ src/PluginManager/include/PluginManager/Module/ModuleIntf.h
