+++ src/PluginManager/src/Module/ModuleManager.cpp
@@ -47,10 +47,9 @@
 #include <Kernel/Enum.h>
 #include <Kernel/QUuidHasher.h>
 #include <Kernel/Sort.h>
-#include "SkyConnectIntf.h"
-#include "ModuleIntf.h"
+#include <Module/ModuleIntf.h>
+#include <Module/ModuleManager.h>
 #include "DefaultModuleImpl.h"
-#include "ModuleManager.h"
 
 namespace
 {
